move utils file to common dsrl_utils
